adds check of all input parameters, adds print of usage file/help file in case of false path parameter, verbosity level set with -v <high,debug,normal>

+++ MP3_Encoder/encoderdef.h
@@ -10,7 +10,7 @@
 
 #define LAME_QUAL_GOOD 5
 
-	const unsigned int version = 0.1;
+	const float version = 0.10;
     const int PCM_SIZE = 8192;
     const int MP3_SIZE = 8192;
